Added compatibility for Chlorine
