+++ src/main/java/com/minecraftabnormals/abnormals_core/core/mixin/client/ClientWorldMixin.java
@@ -42,20 +42,6 @@ public abstract class ClientWorldMixin extends World {
         super(p_i241925_1_, p_i241925_2_, p_i241925_3_, p_i241925_4_, p_i241925_5_, p_i241925_6_, p_i241925_7_);
     }
 
-    @Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/block/Block;animateTick(Lnet/minecraft/block/BlockState;Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;Ljava/util/Random;)V"), method = "animateTick(IIIILjava/util/Random;ZLnet/minecraft/util/math/BlockPos$Mutable;)V")
-    private void animateTick(Block block, BlockState state, World world, BlockPos pos, Random rand) {
-        if (!AnimateTickEvent.onAnimateTick(state, world, pos, rand)) {
-            block.animateTick(state, world, pos, rand);
-        }
-    }
-
-    @Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/fluid/FluidState;animateTick(Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;Ljava/util/Random;)V"), method = "animateTick(IIIILjava/util/Random;ZLnet/minecraft/util/math/BlockPos$Mutable;)V")
-    private void animateFluidTick(FluidState state, World world, BlockPos pos, Random random) {
-        if (!AnimateFluidTickEvent.onAnimateFluidTick(world, pos, state, random)) {
-            ((FluidInvokerMixin) state.getFluid()).callAnimateTick(world, pos, state, random);
-        }
-    }
-
     @Inject(at = @At("HEAD"), method = "getSkyColor", cancellable = true)
     private void getSkyColor(BlockPos blockPos, float partialTicks, CallbackInfoReturnable<Vector3d> info) {
         if (ACConfig.ValuesHolder.isSmoothSkyColorEnabled() && this.field_239131_x_.func_241683_c_() == DimensionRenderInfo.FogType.NORMAL) {
